Delete only on POST and always current user
This only acts on POST requests and blocks the deletion of other users,
so only the own user can be deleted.

+++ delete_user.php
@@ -7,15 +7,18 @@ require_once __DIR__ . '/vendor/autoload.php';
 
 Helper::loadConfiguration();
 Helper::checkLogin();
-Database::connect();
 
-$sql = 'DELETE FROM mitglieder WHERE pilot_id = :pilotid';
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+    Database::connect();
 
-if (Database::query($sql, ['pilotid' => $_POST['pilot_id']]) !== false) {
-    $response = ['success' => true];
-} else {
-    $response = ['error' => 'Error deleting record'];
+    $sql = 'DELETE FROM mitglieder WHERE pilot_id = :pilotid';
+
+    if (Database::query($sql, ['pilotid' => $_SESSION['mitgliederData']['pilot_id']]) !== false) {
+        $response = ['success' => true];
+    } else {
+        $response = ['error' => 'Error deleting record'];
+    }
+
+    header('Content-Type: application/json');
+    echo json_encode($response, JSON_THROW_ON_ERROR);
 }
-
-header('Content-Type: application/json');
-echo json_encode($response, JSON_THROW_ON_ERROR);
